double -> float

+++ raytrace/Math/Coordinate.hpp
@@ -10,31 +10,31 @@ class Coordinate
 public:
 	Coordinate();
 	Coordinate(const Coordinate &c);
-	Coordinate(double x, double y, double z, double w);
-	explicit Coordinate(const ASTVector &astVector, double w);
+	Coordinate(float x, float y, float z, float w);
+	explicit Coordinate(const ASTVector &astVector, float w);
 	
 	Coordinate &operator=(const Coordinate &c);
 
-	double x() const;
-	void setX(double x);
+	float x() const;
+	void setX(float x);
 
-	double y() const;
-	void setY(double y);
+	float y() const;
+	void setY(float y);
 
-	double z() const;
-	void setZ(double z);
+	float z() const;
+	void setZ(float z);
 
-	double w() const;
-	void setW(double w);
+	float w() const;
+	void setW(float w);
 
-	double &at(int i);
-	const double &at(int i) const;
+	float &at(int i);
+	const float &at(int i) const;
 
-	double &operator()(int i);
-	const double &operator()(int i) const;
+	float &operator()(int i);
+	const float &operator()(int i) const;
 
 protected:
-	double mValues[4];
+	float mValues[4];
 };
 
 inline Coordinate::Coordinate()
@@ -42,7 +42,7 @@ inline Coordinate::Coordinate()
 	mValues[0] = mValues[1] = mValues[2] = mValues[3] = 0;
 }
 
-inline Coordinate::Coordinate(double x, double y, double z, double w)
+inline Coordinate::Coordinate(float x, float y, float z, float w)
 {
 	mValues[0] = x;
 	mValues[1] = y;
@@ -50,7 +50,7 @@ inline Coordinate::Coordinate(double x, double y, double z, double w)
 	mValues[3] = w;
 }
 
-inline Coordinate::Coordinate(const ASTVector &astVector, double w)
+inline Coordinate::Coordinate(const ASTVector &astVector, float w)
 {
 	mValues[0] = astVector.x;
 	mValues[1] = astVector.y;
@@ -72,62 +72,62 @@ inline Coordinate &Coordinate::operator=(const Coordinate &c)
 	return *this;
 }
 
-inline double Coordinate::x() const
+inline float Coordinate::x() const
 {
 	return mValues[0];
 }
 
-inline void Coordinate::setX(double x)
+inline void Coordinate::setX(float x)
 {
 	mValues[0] = x;
 }
 
-inline double Coordinate::y() const
+inline float Coordinate::y() const
 {
 	return mValues[1];
 }
 
-inline void Coordinate::setY(double y)
+inline void Coordinate::setY(float y)
 {
 	mValues[1] = y;
 }
 
-inline double Coordinate::z() const
+inline float Coordinate::z() const
 {
 	return mValues[2];
 }
 
-inline void Coordinate::setZ(double z)
+inline void Coordinate::setZ(float z)
 {
 	mValues[2] = z;
 }
 
-inline double Coordinate::w() const
+inline float Coordinate::w() const
 {
 	return mValues[3];
 }
 
-inline void Coordinate::setW(double z)
+inline void Coordinate::setW(float z)
 {
 	mValues[3] = z;
 }
 
-inline double &Coordinate::at(int i)
+inline float &Coordinate::at(int i)
 {
 	return mValues[i];
 }
 
-inline const double &Coordinate::at(int i) const
+inline const float &Coordinate::at(int i) const
 {
 	return mValues[i];
 }
 
-inline double &Coordinate::operator()(int i)
+inline float &Coordinate::operator()(int i)
 {
 	return at(i);
 }
 
-inline const double &Coordinate::operator()(int i) const
+inline const float &Coordinate::operator()(int i) const
 {
 	return at(i);
 }
