+++ raytrace/Object/Primitive/Plane.hpp
@@ -13,7 +13,7 @@ public:
 
 	static Plane *fromAst(AST *ast);
 
-	static Trace::Intersection intersectPlane(const Base *primitive, const Trace::Ray &ray, const Math::Normal &normal, double displacement);
+	static Trace::Intersection intersectPlane(const Base *primitive, const Trace::Ray &ray, const Math::Normal &normal, float displacement);
 
 protected:
 	virtual void doIntersect(const Trace::Ray &ray, std::vector<Trace::Intersection> &intersections) const;
